三次 三级列表分类的功能

+++ qingcheng_service_goods/src/main/java/com/qingcheng/service/impl/CategoryServiceImpl.java
@@ -92,6 +92,14 @@ public class CategoryServiceImpl implements CategoryService {
      * @param id
      */
     public void delete(Integer id) {
+        //判断是否存在下级
+        Example example =new Example(Category.class);
+        Example.Criteria criteria=  example.createCriteria();
+        criteria.andEqualTo("parentId",id);
+        int count= categoryMapper.selectCountByExample(example);
+        if (count>0){
+            throw  new RuntimeException("存在下级分类不能删除");
+        }
         categoryMapper.deleteByPrimaryKey(id);
     }
